requirements.txt
python-telegram-bot
python-telegram-bot[job-queue]
requests
beautifulsoup4
flake8
pytest
pre-commit
validators
selenium
